Solved race on updating Editor registry data in case of multiple registerLanguage calls
In certain cases, not all editors were configured correctly. This was caused by the scheduling of the async task to register the mappings, that at the time of executing this mapping might have been outdated already.

If the outdated update was scheduled after the new one, we would loose languages.